Fix movement issues, where it would move in the opposite directio nthan what was given to it.
When a Creature eats a piece of food, it will then clear its .seeing attribute, which allows it to check for any new food in front and react accordingly.

Creatures now have updated reactions. Reactions have 2 stages now: Move Towards or Away. If it is Towards, it continues moving as it is now. If it is away, then it chooses from the 3 other directions.